Change the prototype of csum_tcpudp_nofold the same way as all other platforms.  This tiny change saves 128 bytes of kernel size ...

+++ include/asm-mips/checksum.h
@@ -124,8 +124,8 @@ static inline unsigned short ip_fast_csum(unsigned char *iph, unsigned int ihl)
  * computes the checksum of the TCP/UDP pseudo-header
  * returns a 16-bit checksum, already complemented
  */
-static inline unsigned int csum_tcpudp_nofold(unsigned int saddr,
-	unsigned int daddr, unsigned short len, unsigned short proto,
+static inline unsigned int csum_tcpudp_nofold(unsigned long saddr,
+	unsigned long daddr, unsigned short len, unsigned short proto,
 	unsigned int sum)
 {
 	__asm__(
